Update Headlights.ino
I set the currentState to SOMETHING_ELSE because the switch wasn't setting it.  Maybe this is due to not actually driving the Analog input.  IDK.
Also, frametime was an INT.  I changed to LONG because ELAPSED is in microseconds.
Also, Logan hit it on the head.  There was a typo error: Pallet vs Palette.  I guess my New Mexico White Trash heritage tends to Pallet, since we used to burn those in the fires down at the Rio Grande dry riverbed.  Haha, <Spock>?